+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## [0.0.1] - 2023-05-05
 
-* TODO: Describe initial release.
+* Login using http server request
+
+## [0.0.2] - 2023-05-05
+
+* Login using http server request